client: Create a new HTTP session for each request
The globally shared connection resets after some time.
When it's reset, the web worker will optimistically try to use it anyway,
and fail with a connection error.

WIth many web workers running behind a load balancer,
this means that a huge portion of login requests would randomly land on a broken session
and it was common for people to need to retry 3-4 times
until their login attempt finally worked.

You can see here how within the sentry project's own code,
sessions are created on demand instead of being shared globally:
https://github.com/getsentry/sentry/blob/27cc0fed47732dec907668a4529ca39bb384bf5a/src/sentry/http.py#L271